Refactor contract-estimation-tests.spec.ts
